ALL backend API except pushservice

+++ api/v1/user.go
@@ -1,7 +1,9 @@
 package v1
 
 import (
+	"ginrss/middleware"
 	"ginrss/model"
+	Redismoon "ginrss/redismoon"
 	"ginrss/utils/errmsg"
 	"github.com/gin-gonic/gin"
 	"net/http"
@@ -14,20 +16,6 @@ var code int
 func AddUser(c *gin.Context) {
 	var data model.User
 	_ = c.ShouldBindJSON(&data)
-
-	//var msg string
-	//var validCode int
-	//msg, validCode = validator.Validate(&data)
-	//if validCode != errmsg.SUCCSE {
-	//	c.JSON(
-	//		http.StatusOK, gin.H{
-	//			"status":  validCode,
-	//			"message": msg,
-	//		},
-	//	)
-	//	c.Abort()
-	//	return
-	//}
 
 	code = model.CheckUser(data.Username)
 	if code == errmsg.SUCCSE {
@@ -45,6 +33,22 @@ func AddUser(c *gin.Context) {
 	)
 }
 //查询用户(1)
+func GetAllUsers(c *gin.Context)  {
+	Users := model.GetAllUsers()
+	count := len(Users)
+	code = errmsg.SUCCSE
+	c.JSON(
+		http.StatusOK, gin.H{
+			"status":  code,
+			"users":  Users,
+			"total" : count,
+			"message": errmsg.GetErrMsg(code),
+		},
+	)
+}
+
+
+
 
 
 //:todo 后台
@@ -52,13 +56,30 @@ func AddUser(c *gin.Context) {
 //编辑用户
 //删除用户
 func DeleteUser(c *gin.Context){
+	tokenClaim, _ := c.Get("tokenUser")
+	tClaim := tokenClaim.(*middleware.MyClaims)
+	userName := tClaim.Username
+	userID := model.SearchUserId(userName)
 	id, _ := strconv.Atoi(c.Param("id"))
-
-	code = model.DeleteUser(id)
-
+	//试图删除自己
+	if userID == uint(id) {
+		code = errmsg.ERROR
+	}else {
+		code = model.DeleteUser(id)
+	}
 	c.JSON(
 		http.StatusOK, gin.H{
 			"states":code,
+			"message":errmsg.GetErrMsg(code),
+		})
+}
+
+func GetActiveUsers(c *gin.Context)  {
+	activeUsers := Redismoon.Getactiveusr()
+	c.JSON(
+		http.StatusOK, gin.H{
+			"states":code,
+			"activeUsers":activeUsers,
 			"message":errmsg.GetErrMsg(code),
 		})
 }
